+++ pytest_adaptavist/_helpers.py
@@ -83,17 +83,20 @@ def get_item_nodeid(item: pytest.Item) -> str:
     return ""
 
 
-def html_row(condition: bool, message: str) -> str:
+def html_row(condition: Literal["passed", "failed", "blocked"], message: str) -> str:
     """Generate an html status row to be displayed in test case results."""
     if not message:
         return ""
 
-    if condition:
+    if condition == "passed":
         background_color = "rgb(58, 187, 75)"
         badge_text = STATUS_PASS
-    else:
+    elif condition == "failed":
         background_color = "rgb(223, 47, 54)"
         badge_text = STATUS_FAIL
+    else:
+        background_color = "rgb(75, 136, 231)"
+        badge_text = STATUS_BLOCKED
 
     return f"<div style='padding: 2pt'><span style='width: auto; margin-right: 4pt; padding: 2pt; border-radius: 4px; background-color: {background_color}; \
             color: white; font-family: monospace; font-size: 10pt; font-weight: bold;'>{badge_text}</span>{message}</div>"