+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## Addedd
+
+* Blocked test cases comments now also get a badge
+
 ## Changed
 
 * The date format in comments is formatted now as %Y-%m-%d %H:%M